id integer

+++ src/main/java/org/iesvdm/tienda/repository/FabricanteRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface FabricanteRepository extends JpaRepository<Fabricante, Long> {
+public interface FabricanteRepository extends JpaRepository<Fabricante, Integer> {
     
 }
